changes made to migrations for deploy

+++ server/db/migrations/20180719102618_pins.js
@@ -2,8 +2,8 @@ exports.up = (knex, Promise) => {
   return knex.schema.createTable('pins', table => {
     table.increments('pin_id').primary()
     table.string('pin_name')
-    table.integer('pin_lat')
-    table.integer('pin_long')
+    table.decimal('pin_lat')
+    table.decimal('pin_long')
     table.integer('area_id')
     table.integer('emotion_type')
     table.string('comment')
